refactor: use toml fie dependencies instead of writing it directly inside the gradle
