+++ app/build.gradle.kts
@@ -56,10 +56,8 @@ dependencies {
     androidTestImplementation(libs.androidx.ui.test.junit4)
     debugImplementation(libs.androidx.ui.tooling)
     debugImplementation(libs.androidx.ui.test.manifest)
-
-    implementation("androidx.compose.material3:material3:1.2.1")
-    implementation("androidx.navigation:navigation-compose:2.7.7")
-
+    //Navigation compose
+    implementation(libs.androidx.navigation.compose)
 }
 
 tasks.register("installGitHooks") {
