+++ app/src/main/java/com/london/tudee/presentation/components/bottom_navigation_bar/Routes.kt
@@ -1,4 +1,4 @@
-package com.london.tudee.presentation.components.BottomNavigationBar
+package com.london.tudee.presentation.components.bottom_navigation_bar
 
 object Routes {
     const val HOME = "home"
